refactor(mapper): extract ResultToTargetGraph to top-level mapper package
Move ResultToTargetGraph and idmapper out of internal/ so external
consumers (e.g. the internal monorepo compute package) can import
them. Proto↔entity conversion helpers remain in internal/mapper.

Also fixes trailing newline lint issue in target_graph.go.
