+++ internal/mapper/BUILD.bazel
@@ -12,11 +12,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//core/errors",
-        "//core/targethasher",
         "//entity",
-        "//internal/mapper/idmapper",
         "//tangopb",
-        "@com_github_bazelbuild_buildtools//build_proto",
         "@org_uber_go_yarpc//encoding/protobuf",
         "@org_uber_go_yarpc//yarpcerrors",
     ],
@@ -32,7 +29,6 @@ go_test(
     embed = [":mapper"],
     deps = [
         "//core/errors",
-        "//core/targethasher",
         "//entity",
         "//tangopb",
         "@com_github_stretchr_testify//assert",
