+++ internal/mapper/target_graph_test.go
@@ -6,7 +6,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/uber/tango/core/targethasher"
 	"github.com/uber/tango/entity"
 	"github.com/uber/tango/tangopb"
 )
@@ -86,15 +85,6 @@ func TestProtoToGetTargetGraphRequest(t *testing.T) {
 			assert.Equal(t, tt.want, got)
 		})
 	}
-}
-
-func TestResultToTargetGraph_EmptyResult(t *testing.T) {
-	t.Parallel()
-
-	targets, meta, err := ResultToTargetGraph(t.Context(), targethasher.Result{})
-	require.NoError(t, err)
-	assert.Empty(t, targets)
-	assert.NotNil(t, meta)
 }
 
 func TestGetTargetGraphResponseToProto_Targets(t *testing.T) {
