+++ mapper/BUILD.bazel
@@ -1,0 +1,25 @@
+load("@rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "mapper",
+    srcs = ["mapper.go"],
+    importpath = "github.com/uber/tango/mapper",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//core/targethasher",
+        "//entity",
+        "//internal/mapper/idmapper",
+        "@com_github_bazelbuild_buildtools//build_proto",
+    ],
+)
+
+go_test(
+    name = "mapper_test",
+    srcs = ["mapper_test.go"],
+    embed = [":mapper"],
+    deps = [
+        "//core/targethasher",
+        "@com_github_stretchr_testify//assert",
+        "@com_github_stretchr_testify//require",
+    ],
+)
